refactor: split gl_device by resource family, slim headers
The single 654-line gl_device.cpp is split into one file per resource
family: gl_buffer.cpp, gl_texture.cpp, gl_shader.cpp, gl_pipeline.cpp.
gl_device.cpp keeps only the device lifecycle, swapchain, command
encoder, and lookup_* accessors. Each per-resource .cpp defines its
slice of gl_device::... members directly, so they retain access to the
private pools without any new accessor surface.

Header bodies move out of headers: the gl_pool template lives in its
own gl_pool.hpp, the GL-side record structs in gl_resources.hpp, and
the inline accessors in renderer.hpp / gl_device.hpp move to their .cpp
files. gl_device.hpp drops from 317 to 125 lines and every backend
.cpp now sits at or below ~220 lines.
